Removed unused code an move the page object methods in a flow file
